feat(v3): widget layer on v3 core
